Fixed compilation error on Linux. xwConfigBase::Read takes a double*, cannot convert from float*
git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@2040 610dfd11-2700-0410-9b68-9ae28eac5983

+++ H3DViewer/src/WxFrame.cpp
@@ -2416,8 +2416,8 @@ void WxFrame::LoadSettings( bool from_config ) {
     int caching_delay;
     string frustum_culling_mode;
     bool use_default_shadows;
-    float default_shadow_darkness;
-    float default_shadow_depth_offset;
+    double default_shadow_darkness;
+    double default_shadow_depth_offset;
     bool prefer_vertex_buffer_object;
     if( from_config ) {
       if (h3dConfig->Exists(wxT("/Settings/GraphicsCaching"))) {
